Push enclave remoting delegation into EnclaveFactory and use service loader to inject EnclaveClient instances to connect to remoted instances.
