+++ enclave/enclave-api/pom.xml
@@ -13,11 +13,20 @@
 
     <dependencies>
 
+        
         <dependency>
             <groupId>com.jpmorgan.quorum</groupId>
-            <artifactId>encryption-api</artifactId>
+            <artifactId>config</artifactId>
         </dependency>
-
+        <dependency>
+            <groupId>com.jpmorgan.quorum</groupId>
+            <artifactId>key-pair-converter</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>${project.groupId}</groupId>
+            <artifactId>encryption-api</artifactId>
+            <version>${project.version}</version>
+        </dependency>
         <dependency>
             <groupId>com.jpmorgan.quorum</groupId>
             <artifactId>shared</artifactId>
