+++ enclave/enclave-api/src/main/java/com/quorum/tessera/enclave/EnclaveClient.java
@@ -1,0 +1,6 @@
+package com.quorum.tessera.enclave;
+
+
+public interface EnclaveClient extends Enclave {
+    
+}
