+++ enclave/enclave-api/src/main/java/com/quorum/tessera/enclave/EnclaveFactory.java
@@ -1,13 +1,40 @@
 package com.quorum.tessera.enclave;
 
-import java.util.ServiceLoader;
+import com.quorum.tessera.ServiceLoaderUtil;
+import com.quorum.tessera.config.AppType;
+import com.quorum.tessera.config.Config;
+import com.quorum.tessera.config.ServerConfig;
+import com.quorum.tessera.config.util.EnvironmentVariableProvider;
+import com.quorum.tessera.encryption.KeyManagerImpl;
+import com.quorum.tessera.encryption.KeyPair;
+import com.quorum.tessera.encryption.PublicKey;
+import com.quorum.tessera.keypairconverter.KeyPairConverter;
+import com.quorum.tessera.nacl.NaclFacadeFactory;
+import java.util.Collection;
+import java.util.Optional;
 
-public interface EnclaveFactory<C> {
+public interface EnclaveFactory {
 
-    Enclave create(C config);
+    default Enclave create(Config config) {
+            Optional<ServerConfig> enclaveServerConfig = config.getServerConfigs().stream()
+                .filter(sc -> sc.getApp() == AppType.ENCLAVE)
+                .findAny();
+
+        if (enclaveServerConfig.isPresent()) {
+            return EnclaveClientFactory.create().create(config);
+        }
+
+        KeyPairConverter keyPairConverter = new KeyPairConverter(config, new EnvironmentVariableProvider());
+        Collection<KeyPair> keys = keyPairConverter.convert(config.getKeys().getKeyData());
+
+        Collection<PublicKey> forwardKeys = com.quorum.tessera.encryption.KeyFactory.convert(config.getAlwaysSendTo());
     
+        return new EnclaveImpl(NaclFacadeFactory.newFactory().create(), new KeyManagerImpl(keys, forwardKeys));
+    }
+
     static EnclaveFactory create() {
-        return ServiceLoader.load(EnclaveFactory.class).iterator().next();
+        return ServiceLoaderUtil.load(EnclaveFactory.class)
+                .orElse(new EnclaveFactory() {});
     }
 
 }
