--- enclave/enclave-api/src/test/java/com/quorum/tessera/enclave/MockEnclaveFactory.java
@@ -1,10 +1,0 @@
-
-package com.quorum.tessera.enclave;
-
-import static org.mockito.Mockito.mock;
-
-public class MockEnclaveFactory implements EnclaveFactory<Object> {
-     public Enclave create(Object config) {
-         return mock(Enclave.class);
-     }
-}
