--- enclave/enclave-api/src/test/resources/META-INF/services/com.quorum.tessera.enclave.EnclaveFactory
@@ -1,1 +1,0 @@
-com.quorum.tessera.enclave.MockEnclaveFactory
