+++ enclave/enclave-jaxrs/src/main/resources/META-INF/services/com.quorum.tessera.enclave.EnclaveClientFactory
@@ -1,0 +1,1 @@
+com.quorum.tessera.enclave.rest.RestfulEnclaveClientFactory
