--- enclave/enclave-jaxrs/src/main/resources/META-INF/services/com.quorum.tessera.encryption.EnclaveFactory
@@ -1,1 +1,0 @@
-com.quorum.tessera.enclave.rest.RestfulEnclaveFactory
